+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Release (2025-MM-DD)
 - `loadbalancer`: [v0.1.0](services/loadbalancer/CHANGELOG.md#v010)
   - Initial onboarding of STACKIT Java SDK for Load balancer service
+- `alb`: [v0.1.0](services/alb/CHANGELOG.md#v010)
+  - Initial onboarding of STACKIT Java SDK for Application load balancer service
 - `objectstorage`: [v0.1.0](services/objectstorage/CHANGELOG.md#v010)
   - Initial onboarding of STACKIT Java SDK for Object storage service
+- `serverupdate`: [v0.1.0](services/serverupdate/CHANGELOG.md#v010)
+  - Initial onboarding of STACKIT Java SDK for Server Update service
 
 ## Release (2025-10-29)
 - `core`: